refactoring in card components
