+++ client/src/components/Carousel/Carousel.js
@@ -193,7 +193,9 @@ const Carousel = ({
 						slidesToShow={slides.length <= 3 ? slides.length : 3}
 					>
 						{slides?.map((slide) => {
+							// eslint-disable-next-line no-console
 							console.log('===============')
+							// eslint-disable-next-line no-console
 							console.log(slide)
 							return (
 								<div key={slide.customId}>
